Revert to using defer on zap logger sync
See https://github.com/uber-go/zap/issues/880 for a known issue in
zap.Sync() that causes the function to always generate an error.
For now, revert to using the defer call on logger.Sync() and instruct
errcheck to ignore the fact that we are not checking the possible error
returned by Sync().