Revert to prior retry logic to get image findings
The getImageFindings function was previously refactored in an attempt to
simplify the retry logic and rely on the WaitUntilImageScanComplete
waiter to fulfill this need. The refactor introduced a bug where
invocations of the Lambda would intermittently fail with a nil pointer
dereference error. In these situations, the initial call to
DescribeImageScanFindings would not return an error, but the
result.ImageScanFindings would be nil, leading to the nil pointer
dereference in the isOldScan function.

The isOldScan function should have better input validation, but that in
and of itself would not solve the intermittent errors as additional
retry logic would be needed in getCurrentImageFindings.

This change is intended to return the relevant code to a known good
state.